+++ apps/expo/src/app/_layout.tsx
@@ -1,19 +1,52 @@
 import { Stack } from "expo-router";
+import * as SecureStore from "expo-secure-store";
 import { StatusBar } from "expo-status-bar";
+import { ClerkProvider } from "@clerk/clerk-expo";
 
 // import { TRPCProvider } from "~/utils/api";
 
 import "../styles.css";
 
+import { Text } from "react-native";
+import Constants from "expo-constants";
 import { useColorScheme } from "nativewind";
+
+const tokenCache = {
+  async getToken(key: string) {
+    try {
+      return SecureStore.getItemAsync(key);
+    } catch (err) {
+      return null;
+    }
+  },
+  async saveToken(key: string, value: string) {
+    try {
+      return SecureStore.setItemAsync(key, value);
+    } catch (err) {
+      return;
+    }
+  },
+};
 
 // This is the main layout of the app
 // It wraps your pages with the providers they need
 export default function RootLayout() {
   const { colorScheme } = useColorScheme();
+  // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+  const clerkPublishableKey = Constants.expoConfig?.extra
+    ?.clerkPublishableKey as string | undefined;
+
+  if (!clerkPublishableKey) {
+    return (
+      <Text>
+        No Clerk Publishable Key found. Please check your environment.
+      </Text>
+    );
+  }
+
   return (
     // <TRPCProvider>
-    <>
+    <ClerkProvider publishableKey={clerkPublishableKey} tokenCache={tokenCache}>
       <Stack
         screenOptions={{
           headerStyle: {
@@ -25,7 +58,7 @@ export default function RootLayout() {
         }}
       />
       <StatusBar />
-    </>
+    </ClerkProvider>
     // </TRPCProvider>
   );
 }
