+++ apps/expo/src/app/new/index.tsx
@@ -1,11 +1,19 @@
 import { Text } from "react-native";
 import { Link } from "expo-router";
+import { SignedIn, SignedOut } from "@clerk/clerk-expo";
+
+import SignInWithOAuth from "~/components/SignInWithOAuth";
 
 export default function Page() {
   return (
     <>
-      <Text>/new page</Text>
-      <Link href="/new/preview">Preview</Link>
+      <SignedOut>
+        <SignInWithOAuth />
+      </SignedOut>
+      <SignedIn>
+        <Text>/new page</Text>
+        <Link href="/new/preview">Preview</Link>
+      </SignedIn>
     </>
   );
 }
