+++ apps/expo/src/app/new/preview/index.tsx
@@ -1,11 +1,19 @@
 import { Text } from "react-native";
 import { Link } from "expo-router";
+import { SignedIn, SignedOut } from "@clerk/clerk-expo";
+
+import SignInWithOAuth from "~/components/SignInWithOAuth";
 
 export default function Page() {
   return (
     <>
-      <Text>/preview page</Text>
-      <Link href="/">Home</Link>
+      <SignedOut>
+        <SignInWithOAuth />
+      </SignedOut>
+      <SignedIn>
+        <Text>/preview page</Text>
+        <Link href="/">Home</Link>
+      </SignedIn>
     </>
   );
 }
